Organize the API routes a bit
