+++ main/api/trpc.ts
@@ -1,0 +1,2 @@
+import { initTRPC } from "@trpc/server";
+export const t = initTRPC.create();
