+++ main/api/window/index.ts
@@ -1,0 +1,25 @@
+import { dialog } from "electron";
+import log from "electron-log";
+
+import { mainWindow } from "../../main-window";
+import { t } from "../trpc";
+
+export const windowRouter = t.router({
+  openDirectoryDialog: t.procedure.query(async () => {
+    const window = await mainWindow.getMainWindow();
+
+    try {
+      const { canceled, filePaths } = await dialog.showOpenDialog(window, {
+        properties: ["openDirectory"],
+      });
+      if (canceled) {
+        return;
+      }
+      return filePaths[0];
+    } catch (e) {
+      log.error(e);
+    }
+
+    return;
+  }),
+});
